改进：使用更多的 pmr
to speed up compile, move member func to .cpp
